# Conflicts:
#	app/src/androidTest/java/com/tosslab/jandi/app/ui/account/AccountHomeActivityTest.java
#	app/src/androidTest/java/com/tosslab/jandi/app/ui/login/login/IntroLoginFragmentTest.java

+++ app/src/main/java/com/tosslab/jandi/app/network/models/ResMessages.java
@@ -428,7 +428,7 @@ public class ResMessages {
         @DatabaseField
         public String ext;
         @DatabaseField
-        public int size;
+        public long size;
 
         @DatabaseField
         public boolean externalShared;
